AuthLogin/requirements.txt
// /requirements.txt

flask
python-dotenv
requests
authlib
six